Add example Akamai OPEN API script
Add an example Python script showing how to call an Akamai OPEN API.

The specific API is /edge-diagnostics/v1/edge-locations, used to fetch
Edge locations.

Parse the Akamai OPEN authentication tokens from a YAML file so that
they're not hard-coded in this repository for security purposes. This
also forms a good baseline for user configuration in the actual
connector script