[DDO-7] Determine next log file
Determine next log file to download and process.

The script keeps track of the last processed log file. It fetches the
available log files from NetStorage. It sorts them by the log start
time (parsed from the filename). It then chooses the log file with
start time after that of the last processed log file. If a log
file was not previously processed, the first log file is selected