Migrate hard-coded values to configuration
There are several values which the user may need to configure.
These include
- Splunk index
- Splunk source type
- LDS timestamp location parse format string
- LDS timestamp strptime format string

These are migrated to the YAML configuration

The format strings are tricky. They're a bit unintuitive, but are
required for the script to support any LDS format. There will
be a section in the readme describing it. It would probably
be smart to hardcode these format strings for some LDS
formats (i.e. dns)