[DDO-9] Tweak log file resume behavior
The script maintains resume data on disk. This is the metadata
of exactly one log file. It contains information like last line processed,
if the file was fully processed, the file name, etc.

The script updates the resume data in the following cases
- A log file was successfully processed. The resume data is updated to
  be the metadata of that file
- A log file was unsuccessfully processed. The resume data is updated to
  be the metadata of that file

The script attempts to resume processing log files using this resume
data.
- If the resume data indicates the log file was fully processed, the
  script resumes just after that log file
- If the resume data indicates the log file was not fully processed,
  the script resumes after the last successfully processed line
  in that log file

The most likely reason for not successfully processing a log is if
Splunk is down/misconfigured. If this happens, multiple log files would
fail processing.

After fixing Splunk, restarting the script should resume at the first
failure.

The script shouldn't update the resume data if no lines were processed.
Otherwise, we'll overwrite the first failure and never process those
lines