[DDO-27] Add Wazuh DNS log decoder and rule
Add the XML definition for a Wazuh decoder that can decode
an LDS log message using DNS format. This decoder applies to
all SysLog messages with program name (in the SysLog header)
equal to "lds_dns_log"

Add the XML definition for a Wazuh rule. This generates an
alert on all messages decoded with the above decoder. Only
alerts are indexed in Wazuh, non-alerting events are discarded.