[DDO-29] Add log delivery to a SysLog server
Refactor the LDS connector script such that logs are sent to an
abstract event handler. This event handler is either a SysLog event
handler or a Splunk event handler.

Implement the Syslog event handler. It uses the Python logging library's
built-in SysLog support with custom formatting to ensure conformity
to the BSD SysLog specification (which is what Wazuh uses)