[DDO-28] Only support delivery to single destination
The LDS connector should only support delivering logs to a single
destination (either Splunk or Wazuh).

We can support delivery to both, but it would require some careful
thought in implementing the resume behavior surrounding batching.
I doubt this is a customer use-case, so let's only allow a single
destination for now